+++ diive/__init__.py
@@ -1,5 +1,4 @@
-from diive.core.plotting.ridgeplot import RidgePlotTS as ridgeplot
-
+from diive.core.plotting.ridgelineplot import RidgeLinePlot as ridgelineplot
 
 # def ridgeplot(*args, **kwargs):
 #     """A convenience function to create and display RidgePlotTS."""